First monomorphize works!

+++ Frontend/FinalAST/src/dev/vale/finalast/instructions.scala
@@ -433,7 +433,7 @@ case class StaticSizedArrayLoadH(
   // owning Car reference element.
   targetOwnership: OwnershipH,
   expectedElementType: ReferenceH[KindH],
-  arraySize: Int,
+  arraySize: Long,
   resultType: ReferenceH[KindH],
 ) extends ExpressionH[KindH] {
   val hash = runtime.ScalaRunTime._hashCode(this); override def hashCode(): Int = hash; override def equals(obj: Any): Boolean = vcurious();
@@ -772,7 +772,7 @@ case class DestroyStaticSizedArrayIntoFunctionH(
   // The prototype for the "__call" function to call on the interface for each element.
   consumerMethod: PrototypeH,
   arrayElementType: ReferenceH[KindH],
-  arraySize: Int
+  arraySize: Long
 ) extends ExpressionH[VoidH] {
   // See BRCOBS, no arguments should be Never.
   // Nevermind, type system guarantees it
